requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0.76
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
PyYAML>=6.0.1
scikit-learn>=1.3.0
scipy>=1.10.0
filterpy>=1.4.5
lapx>=0.5.5
tqdm>=4.66.0
streamlit>=1.33.0
